pyproject.toml
[project]
name = "brasileirao-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.14.0",
    "dbt-duckdb>=1.10.1",
    "duckdb>=1.5.1",
    "fastparquet>=2026.3.0",
    "mem0ai>=1.0.9",
    "openai>=2.30.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
]
